feat: :lock: adding security between accounts in the CRUD of the options

+++ poolplatform/src/main/java/com/poolplatform/features/option/adapters/web/OptionController.java
@@ -6,6 +6,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.poolplatform.adapters.dto.ResponseDTO;
 import com.poolplatform.domain.exceptions.RequestException;
 import com.poolplatform.features.option.adapters.dto.CreateOptionDto;
+import com.poolplatform.features.option.adapters.dto.UpdateOptionDto;
 import com.poolplatform.features.option.domain.OptionService;
 import com.poolplatform.features.option.domain.models.Option;
 import com.poolplatform.features.question.domain.QuestionService;
@@ -16,6 +17,7 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.PutMapping;
@@ -32,10 +34,14 @@ public class OptionController {
     private OptionService optionService;
 
     @PostMapping()
-    public ResponseEntity<ResponseDTO<?>> create(@RequestBody CreateOptionDto optionDto) {
+    public ResponseEntity<ResponseDTO<?>> create(@RequestBody CreateOptionDto optionDto,
+            Authentication authentication) {
         Optional<Question> questionOptional = questionService.get(optionDto.getQuestionId());
         if (questionOptional.isEmpty())
             throw new RequestException("The question does not exist", HttpStatus.NOT_FOUND);
+
+        if (!questionOptional.get().getAuthor().getId().equals((authentication.getPrincipal())))
+            throw new RequestException("No friend", HttpStatus.UNAUTHORIZED);
 
         Option option = new Option();
         option.setOptionText(optionDto.getOptionText());
@@ -50,18 +56,18 @@ public class OptionController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<ResponseDTO<?>> update(@PathVariable String id, @RequestBody CreateOptionDto optionDto) {
-        Optional<Question> questionOptional = questionService.get(optionDto.getQuestionId());
-        if (questionOptional.isEmpty())
-            throw new RequestException("The question does not exist", HttpStatus.NOT_FOUND);
+    public ResponseEntity<ResponseDTO<?>> update(@PathVariable String id, @RequestBody UpdateOptionDto optionDto,
+            Authentication authentication) {
 
         Optional<Option> optionOptional = optionService.get(id);
         if (optionOptional.isEmpty())
             throw new RequestException("The option does not exist", HttpStatus.NOT_FOUND);
 
+        if (!optionOptional.get().getQuestion().getAuthor().getId().equals((authentication.getPrincipal())))
+            throw new RequestException("No friend", HttpStatus.UNAUTHORIZED);
+
         Option option = optionOptional.get();
         option.setOptionText(optionDto.getOptionText());
-        option.setQuestion(questionOptional.get());
 
         optionService.upsert(option);
 
@@ -72,10 +78,13 @@ public class OptionController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<ResponseDTO<?>> delete(@PathVariable String id) {
+    public ResponseEntity<ResponseDTO<?>> delete(@PathVariable String id, Authentication authentication) {
         Optional<Option> optionOptional = optionService.get(id);
         if (optionOptional.isEmpty())
             throw new RequestException("The option does not exist", HttpStatus.NOT_FOUND);
+
+        if (!optionOptional.get().getQuestion().getAuthor().getId().equals(authentication.getPrincipal()))
+            throw new RequestException("No friend", HttpStatus.UNAUTHORIZED);
 
         Option option = optionOptional.get();
         optionService.remove(option);
